Add test and tidy
