+++ xmtp_mls/src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod hash;
+pub mod id;
 #[cfg(test)]
 pub mod test;
 pub mod time;
